+++ src/components/DeletePost.jsx
@@ -1,7 +1,13 @@
 import React from 'react';
 
 const DeletePost = () => {
-  return <div>DeletePost</div>;
+  const deletePost = async targetId => {
+    const { error } = await supabase.from('Post').delete().eq('PostID', targetId);
+    if (error) {
+      console.log('error=>', error);
+    }
+  };
+  return <div><button onClick={() => deletePost()}>포스트 삭제</button></div>;
 };
 
 export default DeletePost;
